fix: move mapequipment methods here
